+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use samplerunity_ShadowMask instead of samplerunity_samplerLightmap for shadow mask
 - Allow to resize reflection probe gizmo's size
 - Improve quality of screen space shadow
+- Remove support of projection model for ScreenSpaceLighting (SSR always use HiZ and refraction always Proxy)
+- Remove all the debug mode from SSR that are obsolete now
 
 ## [4.0.0-preview] - 2018-09-28